reorganize opencl module hierarchy
