+++ pytissueoptics/rayscattering/opencl/config/IPPTable.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 
 class IPPTable:
-    TABLE_PATH = os.path.join(os.path.dirname(__file__), "ipp.json")
+    TABLE_PATH = os.path.join(os.path.dirname(__file__), "../ipp.json")
 
     def __init__(self):
         with open(self.TABLE_PATH, "r") as f:
